Add API to Camera model to flip axes (#7663)
In #7554 / #7488 we changed the default orientation of axes in 3D mode.
This may be undesirable to some users, so this PR adds the ability to
flip them.

It also turns out to be handy when displaying data which has specific
absolute coordinates as well as a common convention for the display
direction, such as geographical data. See the included example which
both demonstrates how to use napari with xarray as well as how to use
this new camera axis-flip API. Currently I'm pretty happy with the API
(see below), but I'm very much open to alternative suggestions!

## API description

This adds an `orientation` attribute to our Camera model. The scene is
assumed to have three axes, z, y, and x, which are aligned with the
depth, vertical, and horizontal axes of the camera. Each of these axes
can then point either:
- away or towards the camera
- down or up
- right or left

I used a StrEnum for each of these values, which makes the API both
robust for type checking and easy to use, e.g.:

```python
viewer.camera.orientation = ('away', 'up', 'right')
```

I also added an `orientation2d` property, because it's strange to set
three values to flip two axes, even though that is the Vispy model.
(camera.flip is a 3-tuple even for 2D cameras in Vispy.) You can see its
use in the example. To show latitude (-90 for south, 90 for north), we
want the vertical axis to point up, not down like in the napari default.
So we do:

```python
viewer.camera.orientation2d = ('up', 'right')
```

and the geographic raster of temperatures is displayed correctly:

![CleanShot 2025-03-03 at 12 42
18@2x](https://github.com/user-attachments/assets/03ee83dd-9272-4991-bf28-524ecd5ff321)

Without this API, the map ends up upside down:

![CleanShot 2025-03-04 at 18 29
50@2x](https://github.com/user-attachments/assets/bb51c197-9fe4-42c8-bd28-687309b46829)

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>