Expand xfail to all macOS CI in 3D screenshot test
Closes #7669

As noted in https://github.com/napari/napari/issues/7669#issuecomment-2702389136,
adding QApplication.processEvents() to the screenshot tests does
not fix the issue, so xfailing should be ok for now -- the API is
still tested in Linux and Win.

This PR can get our CI green while we consider better fixes.