Restyle visibility icon to clarify different interactions compared to clicking layer (#7657)
# References and relevant issues

Implemented to attempt to clarify functionality after discussion in
#7607 and #7648.
cc: @tduvalpm, @psobolewskiPhD, @jni 

# Description

Change qss styling of visibility icon in the layer list. Increases the
clickbox of the visibility icon (which is a nice accessibility
improvement) and _clarifies_ the click area of the visibility icon. Adds
a highlight in the 'visible' state and background in the 'visibility
off' state to visualize highlight the button nature.

The hope is to implicitly help a user identify that clicking on the icon
is _not_ the same as clicking on the layer. Currently, I frequently make
this mistake when browsing through layers quickly, and implicitly
connect clicking the layer icon as _also_ selecting the layer. However,
this clarifies the behavior that clicking on this 'button' will not
select the layer. Changing functionality of click types or separting the
visibility icon to the left of the thumbnail would be a significantly
larger lift as the painting of this icon is currently intrinsically
linked to the rest of the layer elements.

## Final Version


![418825837-40101ffc-2949-43b0-9598-fc2e1062e833](https://github.com/user-attachments/assets/151574b1-7fbd-4527-98cc-9990c51e6b4c)


## Original PR


![image](https://github.com/user-attachments/assets/f223a760-ba3c-439a-a9c0-73944e549759)

## Alternative Implementation 1

Increase padding and shift icon left a bit to improve visual position
```
QtLayerList::indicator {
  width: 16px;
  height: 16px;
  position: absolute;
  left: -2px;
  image: url("theme_{{ id }}:/visibility_off.svg");
  border-radius: 2px;
  padding: 3px;
  background: {{ opacity(highlight, 127) }};
}
```


![image](https://github.com/user-attachments/assets/26262014-f0bb-4294-b697-bb38d6ae61fb)

## Alternative Implementation 2
@jni expressed dislike of this styling, but it is a simple alternative
that we could modify. Clicking _anywhere_ in the visibility square will
toggle, a much larger clickbox than current.

```
QtLayerList::indicator {
  width: 16px;
  height: 16px;
  position: absolute;
  left: -6px;
  image: url("theme_{{ id }}:/visibility_off.svg");
  padding: 7px;
  border-right: 3px solid {{ background }};
}
```


![image](https://github.com/user-attachments/assets/40fd4336-cc1b-465c-9028-361cdd02d596)