Skip `multiple_viewers` example from docs Examples gallery (#7676)
# References and relevant issues
Recent docs crashes, e.g.:
https://github.com/napari/docs/actions/runs/13712919098/job/38379027072?pr=597

# Description
The crashes always happen right after this example with a threading
warning:
`WARNING: QThread: Destroyed while thread is still running`
This is also seen when closing viewers, e.g.
https://github.com/napari/napari/issues/7633
So in this PR I skip this example from the Examples gallery -- it's
rather boring in the gallery anyways.