Fix bounding_box extent for case of 3D multiscale layer (#7545)
# References and relevant issues
Closes: https://github.com/napari/napari/issues/7543

# Description
3D multiscale layers use the lowest resolution when displayed in 3D.
The bounding_box overlay uses the `data_level` of the layer prior to
`ndisplay` change finishing, so if you're looking at a higher rez 2D
slice, then the 3D bounding box will mismatch the displayed data.
In this PR I add a test that fails when the actual vertexes of the
bounding box don't match those that would be expected of the lowest
resolution level.
Then, for the overlay I add a check to see if the layer is 3D and
multiscale, and if so use the lowest resolution bounding box
(augmented).

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>