Show layer status for all visible layers (#7673)
Closes #2955!

This one was a long time coming!

Rather than only showing the status for the active layer, we now show
the status for all visible and loaded layers. The active layer is
indicated by `###`.

I initially skipped all layers behind a translucent layer with
opacity=1.
However, when I was testing it with `example/add_points.py`, it felt
quite jarring to have the image layer disappear from the status just as
we hovered over the point. Sure, it could be remedied by setting opacity
to less than 1, but I thought it's better to just show them all. Happy
to revisit this decision if folks play around with it and find they'd
rather not see hidden layers.

This also allows hovering over *all* layers in a grid mode and seeing
the
right layer info on the status bar, without changing the selection. It's
very nice!

Will upload short video capture shortly.

---------

Co-authored-by: TimMonko <47310455+TimMonko@users.noreply.github.com>
Co-authored-by: Wouter-Michiel Vierdag <michiel.vierdag@embl.de>
Co-authored-by: Carol Willing <carolcode@willingconsulting.com>