Expose new camera orientation API in GUI in ndisplay popup widget (#7686)
# References and relevant issues

Addresses part of #7683. Follow-up to #7663

# Description

I wanted to see how quickly I could mock up and expose the new
`viewer.camera.orientation` API. This is what I have for now, a hopeful
attempt to see if its a desirable solution to #7683.

1. adds widget to camera controls popup to expose
`viewer.camera.orientation` API in both 2D and 3D
2. adds callback between widget values and viewer.camera.orientation,
relying on the enum typing
3. adds tests for the new popup controls


![image](https://github.com/user-attachments/assets/7f75d1bb-d6c6-45a0-9359-3dbae2abbf44)

---------

Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>
Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>