Replace duck-typing check with isinstance check for QRangeSlider (#7712)
In #5757, we added a test for sliders in layer controls that checked for ranged
sliders by checking for the presence of a private attribute `_valuesChanged`.
It's bad practice to check for private attributes in outside libraries.
This instead checks for the correct class in superqt.

See also: https://github.com/pyapp-kit/superqt/pull/283