Add optional font scaling to layer text properties (#7732)
# References and relevant issues
Closes #3185 (for layers managed by TextManager, i.e. not Tracks)

# Description
This PR adds the ability to set `layer.text.scaling = True` (or
equivalent) to scale the text of a layer relative to the size of the
layer and _not_ keeping it as an absolute size with respect to the
canvas. This scaling is useful when there are many points and the text
would otherwise overlap when zoomed out, or become illegible when zoomed
in.

1. Changes `text.size` from PositiveInt to PositiveFloat. This is needed
to prevent jittering of the text during scaling _and_ vispy text is
float compatible. _Note:_ I initially tried to use PositiveFloat from
pydantic, but this does not actually validate correctly. Instead
pydantic.v1 was required for this to work.
2. uses `layer.scale_factor` to calculate font size relative to the
layer, because this scale_factor changes on camera zoom changes, but is
closer to the TextManager than the viewer, since TextManager already
gets passed a layer (but not the vieweR)
3. changes `layer.scale_factor` to be a property that emits an event
4. **changes `VispyCanvas.on_draw()` to use `self.camera.zoom` instead
of `self.viewer.camera.zoom`.** If you investigate the [original PR
description](https://github.com/napari/napari/pull/7732#issuecomment-2746731146)
below, especially Problem Note, you'll see the frustration I was having
with the text size updating at various unexpected events. This change
fixes that issue because `self.viewer.camera` is updated _after_
`on_draw()`, and as such, `on_draw()` was (before this PR) using the
_previous_ zoom level of the camera for its calculations. AFAIK, this
fixes a bug. You can also see the discussion in this zulip thread:
[#reviews > Add optional font scaling to layer text properties
#7732](https://napari.zulipchat.com/#narrow/channel/270578-reviews/topic/Add.20optional.20font.20scaling.20to.20layer.20text.20properties.20.237732)
Ultimately, what allowed me to figure it out what looking at when things
fired in the example in [`64d69b2`
(#7732)](https://github.com/napari/napari/pull/7732/commits/64d69b26e39f55254cde523f63a3faf7df80d7c7)
-- and then was fixed in the next commit.
5. adds `examples/layer_text_scaling.py` -- also useful for testing out
this functionality yourself

Example in action: Magenta is scaled, green is current behavior.

![python_FJl2EFo2Cn](https://github.com/user-attachments/assets/91e2be0a-af4b-4789-a3ae-f7c4e8b92f80)

## Future PRs: 
1. Add text scaling to Tracks Layer (probably, TextManager should be
used here too)
2. Add button in GUI to turn on/off text scaling in appropriate layer
controls widget
3. Consider turning this text scaling ON by default... however this will
'break' visualization that a lot of people currently use.