Fix typo causing mypy error (#7746)
# References and relevant issues
A comment in the tracks layer states in the `colormaps_dict` setter
method:
```
# Ignored type because mypy doesn't recognise colormaps_dict as a property
# TODO: investigate and fix this - not sure why this is the case?
```
I think the issue is just a small typo: `colomaps_dict` instead of
`colormaps_dict`.

# Description
This PR fixes the typo in the name of the setter method.

After correcting the typo and removing the `# type: ignore`, running
`tox -e mypy` throws no errors.

---------

Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>
Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>