Upload failed triangulations as artifacts from benchmarks run (#7766)
# References and relevant issues

For easier debugging of #7747

# Description

In #7632 I have added dump of problematic shapes that crash
triangulation. I think that such an option to debug problematic cases
may be introduced in other part of the codebase.

This PR adds to benchmarks an upload of this data dump for easier
debugging of CI.