Track layer slider max value (#7756)
# References and relevant issues

Closes https://github.com/napari/napari/issues/7728

# Description
This PR takes care of ~~two~~ one thing: 
- In case users set a new `.tail_length` (or `.head_length`) on an
existing tracks layer that is higher than the standard value of 300,
display the track lengths correctly and adjust the new max of the tail
or head length sliders to that new value

The below part was part of the original PR, but I scratched it since
developers decided it would be best to separate it into a new project /
PR in the future.
~~I noticed that the shader tapers off alpha value of the earliest
visible track points (points at the very beginning of tail or end of
head) to 0 (zero). This counteracts the user's input because at a tail
length of 500 for example, the earliest 100-200 points are barely
visible. I clamped the alpha value to 0.5 instead of 0, which I feel is
more reasonable because it leads to a "slide fade" (does not get rid of
the effect completely), but maintains some visibility. One could think
about enabling or disabling this feature with a checkbox.
Below is a direct comparison. Note the change in visibility for blue
(early) track points.~~

```
### Alpha zero (original)
![alpha 0](https://github.com/user-attachments/assets/c5171aba-753d-4878-97a2-c2c493a1344b)

### Alpha clamped to 0.5 (new)
![alpha 5 clamped](https://github.com/user-attachments/assets/784a120a-3260-4a65-a505-59ab864b0590)

```

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>