[bugfix] Adjust scale_bar box scaling and positioning with font size increases (#7750)
# References and relevant issues
Closes https://github.com/napari/napari/issues/7516
Closes https://github.com/napari/napari/issues/7136
(this last one requested a drop shadow, but we agreed the working box is
a better experience)

# Description
This PR does a few things things:
1. it adds a method to ScaleBar visual to recalculate the box height for
a given font size and adjust the layout of the box contexts (the bar and
the text positioning). I use this method when the font size is changed
in the overlay. This method keeps the layout of the box the same
*relative to the bottom of the box* as the font grows the box upwards.
2. Changes the y_size and y_offset defaults in the overlay to actually
match what they do/are: the half-size of the default box and the offset
from the edge. The sum remains the same, so default behavior is
unchanged, but it's easier to reason about the needed changes when the
font and box height change.
3. The above mentioned method is called when the font size of the scale
bar overlay is changed, and for the top positions, the y_offset is
adjusted by the new y_size (half of the box height) such that the top
edge of the box is fixed even as the font increases and the box expands
downwards.
4. updates the tests to check/account for the new behavior.

All of this results in the following, a font_size 30 overlay, at top
right:
With this PR:
<img width="1192" alt="image"
src="https://github.com/user-attachments/assets/cc231800-c2e1-4423-854a-30c26058221e"
/>

With main:
<img width="1192" alt="image"
src="https://github.com/user-attachments/assets/397bf7af-6f66-4f1b-a4d8-7418495c0601"
/>

---------

Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>