Add handedness property to camera and display in GUI (#7770)
# References and relevant issues

Closes #7683 (unless we want to wait until orientation is added to
preferences, which is coming in next PR).

# Description

This was done together with Juan during my introduction to paired
programming; it was a really great experience and I learned so much!. We
sought to add a representation of handedness to the Camera model and
GUI. There is not one single orientation that represents 'right' vs
'left' hand, so this PR allows users to set the orientation as needed
for their data, and then to validate or interpret correctly the
handedness of the camera. This handedness is the most useful property
for comparing data visualization with other software, because handedness
is dependent axes orientation _and_ axes order, which differs commonly.

1. Adds `handedness` property to camera model, with a very clever
implementation by Juan!
2. Adds an updating 'handedness' symbol to the ndisplay / Camera popup
in the GUI. The tooltip also updates with handedness change.
3. Adds tests for both camera property and gui additions.


![python_6u8WhcZAzW](https://github.com/user-attachments/assets/f73898ec-9156-4f73-ab7f-ee2a7cc17fe1)

## Included Fixes:

1. Help symbol is no longer a hardcoded color and now inherits properly
from the theme.
4. A grid layout is used for aligning popup widgets rather than forms.
This will be easier to maintain and allows adding widgets in offset
positions.

## Alternative handedness icon

The icons take inspiration from this sort of visualization of
handedness. (our thumb is Z, since we are ZYX)
https://bevy-cheatbook.github.io/img/handedness.png


![image](https://github.com/user-attachments/assets/4a401e09-1fd2-4fb6-b38d-09749f1ece37)

---------

Co-authored-by: Juan Nunez-Iglesias <juan.nunez-iglesias@monash.edu>
Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>