Add toggle 3D view action (#7721)
Closes #7706.

# Description

This adds a simple "Toggle 2D/3D Camera" menu item, which allows
toggling the camera from the command palette.

# Original message 👇 

Adds a toggle 3D view action with the title "3D View", which appears ticked in
the menu if the view is 3D, and unticked otherwise. This action allows toggling
3D view in the command palette.

One thing I don't like is that we have Cmd-Y as a shortcut for toggling 3D, but
that does not appear in the menu here. I *think* (correct me if I'm wrong
@DragaDoncila) that currently we can't change app-model keyboard shortcuts in
the settings panel. So adding Cmd-Y here would mean losing the configurability
of that shortcut. It might still be worth it because most people aren't aware
of the keyboard shortcuts, and this would increase discoverability?

---------

Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>