Fix layout of help symbols in grid popup (#7771)
# References and relevant issues
During #7770, I use gridlayout instead of formlayout to align widgets in
the popups. This does so for the Grid button popup in the GUI.

# Description
Spacing of help symbols was not properly aligning with popup widgets due
to hardcoded offsets. This is especially apparent as changing the font
size in settings or with high DPI scaling. This change allows for
dynamic alignment of help symbols in the popup widgets

1. Use grid layout to align widgets and help symbols. (spanning rows is
built-in!)
2. Remove hardcoded offsets for spacing help symbols.

Main: (larger font size)

![image](https://github.com/user-attachments/assets/9b13a0c2-2e97-4297-a0e9-c02d6d407be2)

This PR: (larger font size)

![image](https://github.com/user-attachments/assets/68d4109f-e87f-420d-959d-641b0df7113e)