Maintain 3D camera angles when switching between 2D and 3D views (#7765)
# References and relevant issues

Depends on #7742. While I cannot find any specific issues open, I have
observed myself and others dissapointed that `viewer.reset_view()` gets
called in various events, especially when toggling ndisplay and toggling
grid view.

I think this is the perfect change for 0.6.0 as its a major shift
(despite a small change) in how the viewer works for people. I would
really need to be convinced why resetting the camera every time (by
default) is the right way to do things.

# Description

This PR changes toggling viewer state events to use
`viewer.fit_to_view()` instead of `viewer.reset_view()`. In addition,
this PR stores the angles of the 3D camera when switching to 2D, and
restores them upon return to the 3D camera. What this means is that no
longer will you lose your place in the data after a quick toggle.

To replicate the old behavior, toggling ndisplay or grid mode will now
require calling (when in 3D only) `viewer.reset_view()` or pressing the
home button in the GUI.


![python_BQxSKTOAbG](https://github.com/user-attachments/assets/cc758700-88fd-430b-912d-88fdda597dfa)

Added a test that fails on current main and passes on this PR. This also
protects us from accidentally reverting this functionality.

---------

Co-authored-by: Ashley Anderson <aganders3@gmail.com>
Co-authored-by: Carol Willing <carolcode@willingconsulting.com>
Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>
Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>