Revert status bar changes for single active layer (#7775)
This is a partial fix for #7700. This is actually what I had intended to
do in #7673 but didn't get done in time: with this PR, single layer
status messages don't change, and we only have the new "wonky" status
bar with multiple selections or grid mode, which were broken on main
before.

I still want to make some status bar improvements but this was otherwise
holding up 0.6.0, so I'd like to get this in for the next alpha, and
then work on the improvements, which require a bit more surgery in the
Qt status bar display.

@psobolewskiPhD what do you think? With this PR, #7673 becomes a strict
improvement because it doesn't touch the status bar display *unless* we
are in a formerly broken state, that is, with multiple layers selected.
I think, anyway!

So by merging this, we would buy ourselves more time to implement status
bar improvements without holding up the release.

---------

Co-authored-by: Carol Willing <carolcode@willingconsulting.com>