Use proper token in update dependencies workflow (#7821)
# References and relevant issues

Follow up to #7820.

# Description

In #7820, we changed the dependencies update bot to make pull requests
from its own fork, rather than by pushing to the main napari repo.
However, the token used in that step had permissions for the main repo
and not for the fork. This PR updates the token to use the one for
commenting on the issue, which should have the correct permissions.