Fix issues displaying polygons with holes in Shapes (#6654)
# References

~~Partial fix for~~ Closes #5673. (!!!)
Closes #1653. 🥳 
Closes #1058. 🥳 
Closes #7799.

Depends on #7622 

# Description

(Updated 29/2/2024.)
(Updated 16/4/2025.)

This PR adds three preprocessing steps to polygon data during meshing in
Shapes layers:
- discard duplicate polygon vertices, and recast edges in terms of
unique vertices.
- discard duplicate edges — these are found when going to interior
polygons depicting holes and when coming back.
- remove triangles from the triangulation that are not in the polygon.
(only for the `triangle` backend, which otherwise includes these
triangles.)

For the VisPy step, we go directly into the Triangulation class, because
the PolygonData class completely ignores input edges when building a
triangulation.

## Current limitations:

- ~~only works when `triangle` is installed (`py-triangle` in
conda-forge)~~ **Edit:** nope, works with plain VisPy too!
- ~~currently on macOS, I am getting random segfaults when calling
triangle.triangulate. I don't know what triggers them, but I suspect
we're just going to have to implement our own triangulation, maybe with
Numba.~~ **Edit:** these were caused by repeated vertices, against which
triangulation algorithms are generally not robust. Uniquifying the
vertices fixed the problem!
- ~~I still need to filter out the edges of the polygon for drawing, and
maybe also for highlighting. Currently, the edges connecting the
interior and exterior polygons are visible, which looks crap.~~
**Edit:** this is now done!
- ~~the polygons are processed twice, once for the interiors and once
for the edges.~~ **Edit:** fixed by @Czaki 🥳
- ~~the processing is probably a fair bit of overhead when many datasets
would get no benefit. It might be worthwhile adding a flag to turn these
checks on/off.~~ **Edit:** fixed by @Czaki 🥳

## Summary of the issue and strategies:

It turns out that what we actually need is a *constrained Delaunay
triangulation*[1], that is, a Delaunay triangulation that preserves
specific edges from the vertices. That’s what the 'p' option in
`triangle.triangulate` does *but* it doesn’t work if the edges are
actually not planar — that’s why the triangulation was failing for this
polygon even if triangle was installed.

This PR fixes that by finding edges that appear twice in the polygon and
removing them.

Additionally, certain triangles from the resulting triangulation need to
be removed, and this is easy to do by running
`skimage.measure.points_in_poly` on the triangle centers, which are fast
to compute with NumPy indexing.

[1]: http://en.wikipedia.org/wiki/Constrained_Delaunay_triangulation

---------

Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>
Co-authored-by: Wouter-Michiel Vierdag <w-mv@hotmail.com>
Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Tim Monko <timmonko@gmail.com>