Add PR dependency workflow (#7828)
# References and relevant issues
napari/docs#124

# Description
Add a CI check whenever phrases like `depends on #...` are used, which
blocks merging until the tagger PR/issue is merged/closed. Works
cross-repo.

A while back we added this to the docs repo, and I think it proved
useful over there.

At the time we briefly discussed adding it here as well, but never got
around to it. What does everyone think?

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>