Add a custom log handler and GUI viewer with filters (#6900)
# References and relevant issues
Followup on
https://github.com/napari/napari/pull/6849#issuecomment-2095676630

# Description
The idea is to have our own custom handler for log messages so we can
then expose the logs as we prefer from the GUI. This will make it easier
for users to copy-paste logs even if they didn't manually open from the
console, and also easier for developers to quickly check things.

I never really worked with the logging module, so I have a feeling that
this is garbage... but oh well, let's see how bad :P

To test, open napari and do some stuff. Then, from the terminal or
console run:
```py
napari._LOG_STREAM.logs_at_level(10)
```

EDIT: Added a very quick-and-dirty GUI for the logger based on the
implementation of the `About` widget. For how bad it is, it's actually
ok xD You can open it from the menu `Help -> Show log".

cc @czaki

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>
Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>
Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>
Co-authored-by: Lucy Liu <jliu176@gmail.com>
Co-authored-by: Carol Willing <carolcode@willingconsulting.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>