[bugfix] ensure erasing and fill work with swap and preserve labels (#7816)
# References and relevant issues
Closes: https://github.com/napari/napari/issues/7812

# Description
This PR attempts to make the behavior of paint and fill work with
`swap_selected_and_background_labels` (`X` keybind) and preserve labels:
- it enables erasing the last used label with *both* paint/fill when
swap has been used to switch to background, rather than returning, but
checking the old label vs. the _prev_selected_label
e.g. on main, the following happens:
- paint a label, set preserve, hit swap binding, paint over to erase ->
nada, nothing happens
- paint a label, set preserve, hit swap binding, fill over to erase ->
nada, nothing happens

with this Pr:
- in both cases the label that was being painted (_prev_selected_label)
will be erased

To make the behavior more predictable, e.g. when using the
`selected_label` setter/spinbox vs. the `X` keybind, I have the setter
only store `self._prev_selected_label` when the label is set to the
background, rather than all the time. This also prevents confusing
issues with un-expected fills or erases that could occur when
selected_label and _prev_selected_label are different and neither is
None.

I also added two tests that fail on main, as well as a test for the
_prev_selected_label setting behavior.