Update CITATION.cff for 0.6.0 (#7860)
Our CITATION.cff is of course again out of date.

I've moved Lucy to the core devs section (which is in some weird order
but fixing that is out of scope for this PR), and I've added Tim
(@TimMonko could you please check I have your info right?).

@cnstt ~~@hanjinliu @horsto @sfmig @willingc @yarikoptic~~, you have
contributed to the latest napari release, if you would like to appear in
our Zenodo record, please reply here with the following info in a code
block like so:

```
- given-names:
  family-names:
  affiliation:
  orcid: (optional)
  alias: (your GH username)
```

and I'll add it to the file. Thank you!