Gallery: use colors from the South African flag to draw SA (#7859)
For the social media posts about 0.6.0rc0, I found that the dev
example with the polygon-with-holes aka South Africa was
pretty bland with the gray colors. So I picked some colors from
the South African flag to draw it and I think it looks much nicer!

I could of course just leave this in my local stash but it looked
nice so I thought I'd share more widely. 😊


![napari-za](https://github.com/user-attachments/assets/f6028474-7a14-4aad-892b-d0b89241ef95)