[Maint] Update dockerfile for napari 0.6.0 (#7872)
# References and relevant issues
Closes: https://github.com/napari/napari/issues/7847

# Description
This PR does the following:
1. uses the official Python 3.11 base image -- I had trouble with python
3.11 in Ubuntu 22.04 (only a release candidate is available). Ubuntu
24.04 has python 3.12, which I guess could be ok? but it had issues with
pip/pipx and extra warnings from xpra. So I figured the official Python
image was the easiest. Then when we want to update to 3.12 we just
change the version at the top and we don't need to mess round with apt
install changes.
3. Sorts the apt installs to match the Talley setup-qt-libs README, to
make it easier to check that it's correct. Also updated the libegl1 and
libglx-mesa0 installs. Not strictly needed yet, but will be for newer
Linux distros, see:
https://github.com/tlambert03/setup-qt-libs/issues/65 ~~Finally, I
removed `libglib2.0-0` because it's not recommended to be installed
anymore.~~
4. uses the constraints for the pip install to ensure the container is
what passes tests.
5. I tried to get the workflow to fire when the Dockerfile is updated...

I tested locally and it build and ran correctly. Hopefully the workflow
fires...
I will followup with some additional changes to the xpra part, but
that's not needed for 0.6.0