Add inheritance of spatial data for functional plugin that return layer data.  (#6986)
# References and relevant issues

This is part of the implementation of units in napari 

# Description

As we support functional plugins that get and return only layer data, we
should inherit spatial properties from source layers.

In the current implementation, layer with dimensionality lower than
produced data are ignored.

If spatial information of all source data of dimensionality at least the
same as result data is consistent, the new created layer has the same
spatial information as the source layers.

Otherwise, it preserves previous behavior and left default values.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Draga Doncila Pop <17995243+DragaDoncila@users.noreply.github.com>