Do not expose vispy BaseColormaps (#7858)
# References and relevant issues
- fixes #7855

# Description
Some colormap from vispy are broken (see issue for explanation). This PR
simply removes them from the available colormaps.

Updating them in vispy is another option, but I think it's worse because
these colormaps have been in vispy a long time, and people might depend
on their quirks. Also, `hot` is provided by matplotlib and `grays` is
essentially a duplicate of matplotlib's `gray`, so we really only lose
`ice`.

Note that #7846 additionally introduces a `HiLo` colormap, which also
currently cannot be represented by our `Colormap` model. Support for
that will need to be added in that PR.