Surface original error when a selected plugin fails to read file. (#7901)
# References and relevant issues
Discussed in #7826

# Description

When a plugin is selected to read a file, but it fails OR when there is
only a single plugin available to read a file, we raise our own error
message `from` the original error. However, this means the (likely more
useful) original error thrown by the plugin doesn't appear unless you
scroll up in the stack trace, even if the error would be relatively easy
to fix e.g. just install `imagecodecs` or something.

This PR improves our own error message so that the original error is
printed to the user, and they are advised to scroll up to see the full
stack trace. @brisvag and I discussed maybe bringing the actual stack
trace down, but decided that would be a lot more complex, and
potentially confusing for developers who **are** used to just scrolling
up to get to their own code.

Prior to this PR, if you don't have `tifffile` installed (for some
reason) and try to open a tiff file:


![image](https://github.com/user-attachments/assets/a8c06009-5e48-43ad-b6f3-7b0147db70ff)

With this PR:


![image](https://github.com/user-attachments/assets/7ac2d792-48da-49e3-bd48-523758df8fcb)

---------

Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>
Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>
Co-authored-by: Draga Doncila <ddon0001@student.monash.edu>