Fix handle mouse events (#7936)
# References and relevant issues
- fix #7931
- updates the overlay dev example to showcase how to use `event.handled`

# Description
Boy this was a rabbithole! In the end the solution was simple, but
debugging was confusing. It turns out, we never transfered the `handled`
attribute from our napari events to the vispy events, so vispy happily
continued to call the callbacks.

You might think: but wait, doesn't our points callback for `add` use
`handled`? And that works!


https://github.com/napari/napari/blob/c6c71329431b942a8e9379acc7cebda741640465/napari/layers/points/_points_mouse_bindings.py#L123-L134

Well... it doesn't actually work, somehow we thought it did xD

So great, this PR fixes it right? Well, actually not quite, cause even
after the fix, there's a second issue, caused by the fact that we
throttle canvas mouse move callbacks on the napari side:


https://github.com/napari/napari/blob/c6c71329431b942a8e9379acc7cebda741640465/napari/_vispy/canvas.py#L153-L155

However, only napari callbacks are throttled, not vispy ones... this,
combined with the fact that our `add()` callback only declares the event
as handled *in the while loop*, means that if you move your mouse fast
enough, you'll still be able to move the canvas. You can test this by
setting the drag threshold to a much higher number (like 1000) here:


https://github.com/napari/napari/blob/c6c71329431b942a8e9379acc7cebda741640465/napari/layers/points/_points_mouse_bindings.py#L120

This should make it impossible to ever drag the canvas in add mode, but
if you move the mouse fast enough it still works.

So I thought: let's mark the event as `handled` before the first
`yield`, and then set `handled = False` if the threshold is surpassed.
However, this doesn't work, because if the first event is `handled` then
the vispy camera never begins its drag callback, and will never restart
it again. Also, we cannot change the `event.type` halfway through the
callbacks, so that's also out of discussion.

----

So, in conclusion:

- this PR fixes the general issue of `event.handled` being ignored by
vispy (fixes #7931)
- I have currently have no real solution for our points `add` event.
However, it should be fine unless a user tries to add a point while
making micromovements at super high speed :P