Don't add widgets to non-contributable menus (#7926)
# References and relevant issues
N/A

# Description
This PR makes sure that widgets can only be added to contributable
menus, rather than any valid napari menu.

Context: In #7262 we tried adding the shapes measuring action to the
layerlist context menu, but this didn't work because the menu is not
contributable. However, in #7877 adding the features table widget to the
layerlist context menu worked fine! This is because we were never
filtering menus for widgets to only include the contributable ones, but
the shapes measuring action is **not** a widget, just a plain command,
and that gets handled in a separate function. This PR fixes that.

Co-authored-by: Draga Doncila <ddon0001@student.monash.edu>