Refactor layer overlays visuals from VispyLayer to VispyCanvas (#7835)
# References and relevant issues
Implemented as part of #7832 and split out into smaller PRs.
Continued in #7836.

# Description
This PR stems from 2 main issues:
1. Layer overlays (for now mainly bounding box, but with #7832 also
colormap) are currently a strange hybrid in in the vispy backend.
Instead of getting instantiated by the `VispyCanvas` (like every other
visual), they are instead handled by the `VispyBaseLayer` class. This
seemed intuitive in the past, but has created lots of issues,
specifically with the ability to communicate across the vispy scenegraph
(for example, look at the ugly node reparenting code that I removed from
this class in this PR).
2. The above issue became too much when I attempted to implement tiling
for our canvas overlays (i.e: the ability have multiple arbitrary
overlays at the top right without them overlapping). For this to work,
they needed to communicate across overlays, without having any real
connection aside from some weird backward event chain that made working
with it tedious and error prone.

So I decided to try to move all the overlay handlding to the
`VispyCanvas` class. This is now a centralized hub for vispy visuals,
where events can be hooked up and removed as approprate, and which is
aware of all elements on the canvas and in the scene. This refactor
shouldn't change any behaviour.

The changes that introduce the actual tiling are in a followup PR
(#7836).

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>